[Refactor : back] Request 이름 변경
